Fixed grammatical error
I noticed a grammatical error and took a quick pass at editing it. No need to use this version, but you may want to take a look at the structure of this sentence and adjust ("with xyz... and is mobile compatible") 

So excited for the new launch! The site looks incredible!